+++ needls/server.py
@@ -494,17 +494,35 @@ def update_settings(ls, *args):
     ls.show_message_log(f"Needs file: {needs_file}")
     try:
         ls.needs_store.set_docs_root(docs_root)
-        ls.needs_store.set_declared_types()
-        ls.show_message_log(f"Declared need types: {ls.needs_store.declared_types}")
-    except ValueError:
+    except Exception as e:
+        ls.show_message_log(f"Something is wrong with Docs root: {docs_root} -> {e}")
         ls.show_message(
             "Error setting document root! Are your settings correct?",
             msg_type=MessageType.Error,
         )
         return
+
+    try:
+        ls.needs_store.set_declared_types()
+        ls.show_message_log(f"Declared need types: {ls.needs_store.declared_types}")
+    except Exception as e:
+        ls.show_message_log(
+            f"Something is wrong with declared need types: {ls.needs_store.declared_types} -> {e}"
+        )
+        ls.show_message(
+            "Error loading declared needs_types from conf.py",
+            msg_type=MessageType.Error,
+        )
+        return
+
     try:
         ls.needs_store.load_needs(needs_file)
-    except ValueError:
+        # for debugging
+        ls.show_message_log(
+            f"Loaded {len(ls.needs_store.needs)} needs from needs.json: {list(ls.needs_store.needs.keys())}"
+        )
+    except Exception as e:
+        ls.show_message_log(f"Failed to load needs.json: {needs_file} -> {e}")
         ls.show_message(
             "Error loading needs.json! Are your settings correct?",
             msg_type=MessageType.Error,